+++ src/StardewModdingAPI.Installer/Enums/ScriptAction.cs
@@ -1,0 +1,12 @@
+﻿namespace StardewModdingApi.Installer.Enums
+{
+    /// <summary>The action to perform.</summary>
+    internal enum ScriptAction
+    {
+        /// <summary>Install SMAPI to the game directory.</summary>
+        Install,
+
+        /// <summary>Remove SMAPI from the game directory.</summary>
+        Uninstall
+    }
+}