+++ src/StardewModdingAPI.Installer/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("StardewModdingAPI.Installer")]
+[assembly: AssemblyProduct("StardewModdingAPI.Installer")]
+[assembly: ComVisible(false)]
+[assembly: Guid("443ddf81-6aaf-420a-a610-3459f37e5575")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]